Add item properties to a project

Add, change or define item properties in a dexter project

add_item_properties(db, item_properties = NULL, default_values = NULL)

Arguments

db
a connection to a dexter database, e.g. the output of start_new_project or open_project

item_properties
A data frame containing a column item_id (matching item_id's already defined in the project) and 1 or more other columns with item properties (e.g. item_type, subject)

default_values
a list where the names are item_properties and the values are defaults. The defaults will be used wherever the item property is unknown.

Value

nothing

Details

When entering response data in the form of a rectangular person x item table, it is easy to provide person properties but practically impossible to provide item properties. This function provides a possibility to do so.

Note that is is not possible to add new items with this function, use touch_rules if you want to add new items to your project.

See also

Examples