extract enorm item parameters

extract enorm item parameters

# S3 method for prms
coef(object, hpd = 0.95, what = c("items", "var", "posterior"), ...)

Arguments

object
an enorm parameters object, generated by the function fit_enorm

hpd
width of Bayesian highest posterior density interval around mean_beta, value must be between 0 and 1, default is 0.95

what
which coefficients to return. Defaults to `items` (the item parameters). Can also be `var` for the variance-covariance matrix (CML only) or `posterior` for all draws of the item parameters (Bayes only)

...
further arguments to coef are ignored

Value

Depends on the calibration method and the value of 'what'. For 'items'#'

CML

a data.frame with columns: item_id, item_score, beta, SE_beta

Bayes

a data.frame with columns: item_id, item_score, mean_beta, SD_beta, <hpd_b_left>, <hpd_b_right>

The posterior distribution and variance covariance matrix are returned as matrices.