Information about the design

This function is useful to inspect incomplete designs

design_info(dataSrc, predicate = NULL)

Arguments

dataSrc
a connection to a dexter database, a matrix, or a data.frame with columns: person_id, item_id, item_score

predicate
An optional expression to subset data, if NULL all data is used

Value

a list with the following components

design

a data.frame with columns booklet_id, item_id, item_position, n_persons

connected_booklets

a data.frame with columns booklet_id, group; booklets with the same `group` are connected to each other.

connected

TRUE/FALSE indicating whether the design is connected or not

testlets

a data.frame with columns item_id and testlet; items within the same testlet always occur together in a booklet

adj_matrix

list of two adjacency matrices: *weighted_by_items* and *weighted_by_persons*; These matrices can be useful in visually inspecting the design using a package like *igraph*