Fit the extended nominal response model

Fits an Extended NOminal Response Model (ENORM) using conditional maximum likelihood (CML) or a Gibbs sampler for Bayesian estimation.

fit_enorm(
  dataSrc,
  predicate = NULL,
  fixed_params = NULL,
  method = c("CML", "Bayes"),
  nDraws = 1000,
  merge_within_persons = FALSE
)

Arguments

dataSrc

a connection to a dexter database, a matrix, or a data.frame with columns: person_id, item_id, item_score

predicate

An optional expression to subset data, if NULL all data is used

fixed_params

Optionally, an enorm object from a previous analysis or a data.frame with parameters, see details.

method

If CML, the estimation method will be Conditional Maximum Likelihood; otherwise, a Gibbs sampler will be used to produce a sample from the posterior

nDraws

Number of Gibbs samples when estimation method is Bayes.

merge_within_persons

whether to merge different booklets administered to the same person, enabling linking over persons as well as booklets.

Value

An object of type enorm. The following methods are supported:

coef

plot

logLik

In addition, many dexter functions accept an enorm object as input, e.g.

ability

plausible_values

plausible_scores

expected_score

Details

The eNRM is a generalization of the PCM and the OPLM. It reduces to the Rasch model for dichotomous items when all itemscores are 0 or 1, is equal to the PCM for polytomous items if all itemscores up to the maximum score occur. It is equal to the oplm if all itemscores have an equal common divisor larger than 1.

To support some flexibility in fixing parameters, fixed_params can be a dexter enorm object or a data.frame. If it is a data.frame, it should contain the columns item_id, item_score and a difficulty parameter beta

References

Maris, G., Bechger, T.M. and San-Martin, E. (2015) A Gibbs sampler for the (extended) marginal Rasch model. Psychometrika. 80(4), 859-879.

Koops, J. and Bechger, T.M. and Maris, G. (2024); Bayesian inference for multistage and other incomplete designs. In Research for Practical Issues and Solutions in Computerized Multistage Testing. Routledge, London.