Variables that are defined in the project

Inspect the variables defined in your dexter project and their datatypes

get_variables(db)

Arguments

db
a dexter project database

Value

a data.frame with name and type of the variables defined in your dexter project

Details

The variables in Dexter consist of the item properties and person properties you specified and a number of reserved variables that are automatically defined like response and booklet_id.

Variables in Dexter are most useful when used in predicate expressions. A number of functions can take a dataSrc argument and an optional predicate. Predicates are a concise and flexible way to filter data for the different psychometric functions in Dexter.

The variables can also be used to retrieve data in get_responses