Latent correlations

Estimates correlations between latent traits. Use an item_property to distinguish the different scales. This function uses plausible values so results may differ slightly between calls. Note: this is a new and slightly experimental function and therefore still a bit slow. It will probably become faster in future versions of dexter.

latent_cor(
  dataSrc,
  item_property,
  predicate = NULL,
  nDraws = 500,
  hpd = 0.95,
  use = "complete.obs"
)

Arguments

dataSrc

A connection to a dexter database or a data.frame with columns: person_id, item_id, item_score and the item_property

item_property

The name of the item property used to define the domains. If dataSrc is a dexter db then the item_property must match a known item property. If datasrc is a data.frame, item_property must be equal to one of its column names.

predicate

An optional expression to subset data, if NULL all data is used

nDraws

Number of draws for plausible values

hpd

width of Bayesian highest posterior density interval around the correlations, value must be between 0 and 1.

use

Only complete.obs at this time. Respondents who don't have a score for one or more scales are removed.

Value

List containing a estimated correlation matrix, the corresponding standard deviations, and the lower and upper limits of the highest posterior density interval