A plot method for the interaction model

Plot the item-total regressions fit by the interaction (or Rasch) model

# S3 method for rim
plot(
  x,
  items = NULL,
  summate = TRUE,
  overlay = FALSE,
  curtains = 10,
  show.observed = TRUE,
  ...
)

Arguments

x

An object produced by function fit_inter

items

The items to plot (item_id's). If NULL, all items will be plotted

summate

If FALSE, regressions for polytomous items will be shown for each response option separately; default is TRUE.

overlay

If TRUE and more than one item is specified, there will be two plots, one for the Rasch model and the other for the interaction model, with all items overlayed; otherwise, one plot for each item with the two models overlayed. Ignored if summate is FALSE. Default is FALSE

curtains

100*the tail probability of the sum scores to be shaded. Default is 10. Set to 0 to have no curtains shown at all.

show.observed

If TRUE, the observed proportion correct at each sum score will be shown as dots. Default is FALSE.

...

Any additional plotting parameters.

Details

Customization of title and subtitle can be done by using the arguments main and sub. These arguments can contain references to the variables item_id (if overlay=FALSE) or model (if overlay=TRUE) by prefixing them with a dollar sign, e.g. plot(m, main='item: $item_id')